update category and checkout page

+++ ecommerce/products/models.py
@@ -10,7 +10,7 @@ class Category(models.Model):
 
 class Product(models.Model):
     product_name=models.CharField(max_length=100)
-    product_price=models.FloatField()
+    product_price=models.DecimalField(max_digits=10, decimal_places=2)
     stock=models.IntegerField()
     product_description=models.TextField()
     product_image=models.FileField(upload_to='static/uploads',null=True)
